+++ js-miniapp-sdk/src/modules/utils.ts
@@ -43,6 +43,11 @@ export interface MiniAppUtilsProvider {
    * @param analyticsInfo Analytics info
    */
   sendAnalytics(analytics: MAAnalyticsInfo): Promise<string>;
+
+  /**
+   * Interface to get list of features supported by the SDK and Host
+   */
+  getFeatureList(): Promise<string[]>;
 }
 
 /** @internal */
@@ -67,5 +72,8 @@ export class MiniAppUtils implements MiniAppUtilsProvider {
       return getBridge().sendAnalytics(analytics);
     }
     return Promise.reject('sendAnalytics Error');
+  }
+  getFeatureList(): Promise<string[]> {
+    return getBridge().getFeatureList();
   }
 }